requirements.txt
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tslearn>=0.6.0
PyYAML>=6.0
scipy>=1.10.0
aeon>=0.7.0
